Rename app pages
